wmMoonClock: fix build with ld --as-needed
I'm forwarding a patch from Michael Bienia <geser@ubuntu.com> that
has been applied to the wmmoonclock package in Ubuntu.

It fixes a build issue that is caused by specifiying libraries before the object
files in the linker command line.
This causes a build failure as Ubuntu's ld defaults to --as-needed, see:
http://wiki.debian.org/ToolChain/DSOLinking#Only_link_with_needed_libraries

[Patch sent by: Felix Geyer <debfx@ubuntu.com>]