Width tags clean
Removed the body and header tags of these pages, because when are joined
by php the final document will have three times the body and the header
tags included.

Removed the width tags, to avoid the space in the right side and the
horizontal slide.

Removed the homewmaker2.png link image, because this text can be made
using formatted text.

Removed a lot of \n \r.