Removed all autogenerated autotools and imake stuff
There's no need to keep autogenerated files in
repository.  Those dockapps which use autotoold must be
autoreconf'ed prior to build, those which use imake must be
xmkmf'ed.  As of wmbiff, AC_PATH_XTRA_CORRECTED was a hack to
avoid bug in older autoconf version, which is no longer needed
(it's 2012 now already).

Signed-off-by: Alexey I. Froloff <raorn@raorn.name>