wmbutton: Middle button enabled by default
The mouse middle button should be enabled by default (see the help
and the manpage). This patch solves this bug.

The MIDMOUSE definition is removed because the value of MIDMOUSE
changes the middle button behaviour. Now, the middle button always works,
except if the user uses the -m argument.

Now the manpage and the help show the same info.