wmbutton: Makefile install rewritten
The current Makefile script doesn't work properly, because the path
for /etc doesn't use DESTDIR and the installation path for $(DESTDIR)/usr/bin
is not created before installing the binary file in the folder.

LDFLAGS are needed in some distros to build the package, for example Debian.

This patch solves these problems.