wmtime: Fixed Debian bug #639626. Applied a patch by Andrew Deason.
Package: wmtime
Version: 1.0b2-10
Tags: patch

wmtime draws a small amount of garbage to the screen when -noseconds is
given, because DrawTime reads beyond the given string buffer. This
doesn't trigger all the time for me (I assume due to the nondeterminism
of reading uninitialized memory or something), but it pretty clearly
happens if I switch VTs.

The attached patch limits the amount of the formatted string we try to
read, so we don't read beyond the end of the string. Fixes it for me.

--
Andrew Deason
adeason@dson.org