Fixed broken links in the FAQ.
Anh K. Huynh brought to the attention of the wmaker-dev mailing list that the links in the FAQ on the Window Maker website were broken.

An archived version of the Window Maker website with a complete FAQ was found on archive.org, and the missing questions and answers were copied and pasted into the current FAQ file.  Some minor formatting occurred to bring the FAQ in line with the website's current style.