dock.php, header.php: Update include files to use absolute paths.
The links and images used in the dock.php header.php include files use relative
paths.  This poses a problem if someone were to include them from a file in a
subdirectory.  In particular, the links and images used by the include files
would be assumed to be in the same subdirectory.

This patch fixes this issue by using absolute paths instead.