Add dockapps section.
This patch essentially adds a mini version of the old dockapps.windowmaker.org
to the main webpage for the dockapps which are in the git repository.

Instead of hosting the tarballs, links to git snapshots on repo.or.cz are given.

The images and descriptions came mostly from archive.org's copy of
dockapps.windowmaker.org.  They are stored in a text-based database
(dockapps.db), which is generated locally by a perl script that pulls the
version information from git.