wmitime: Improve locale support.
Previously, wmitime only had support for English, French, and (in Debian
only) Hungarian.  In addition, the choice was made at compile time.

This patch adds run-time support for any language using the Latin alphabet.
The locale is determined by the user's environment or may be specified on
the command line with the "-l" option.  Note that users whose environment
specifies a non-Latin locale may wish to use, e.g., "-l C", as otherwise
no date will appear.

Note that, for simplicity, the month and day are now displayed as "01 JAN"
as opposed to "JAN 01".  (Previously, the former format was used for English
and the latter for French.)