New library wmgeneral
This patch includes the library wmgeneral. The wmgeneral library
is used in a lot of dockapps:

$ find . -name wmgeneral
./wmkeys/wmgeneral
./wmtime/wmgeneral
./wmtz/wmgeneral
./wmsm.app/wmgeneral
./wmcpufreq/wmgeneral
./wmtv/src/wmgeneral
./wmSMPmon/wmgeneral
./wmckgmail/wmgeneral
./wmifs/wmgeneral
./wmbiff/wmgeneral
./wmfsm/wmgeneral
./wmsun/wmgeneral
./wmmon/wmgeneral
./wmitime/wmgeneral
./wmppp.app/wmgeneral

So, any change in the wmgeneral needs change a lot of dockapps.

I used the wmmon dockapp to move the wmgeneral and then I adjust the
wmmon to use the library. The library must be installed in the system
to build wmmon (wmmon uses the .h files) and is needed to run wmmon
(wmmon uses the libwmgeneral.so shared library).

I did not include any Changelog/README/... file to announce the change
in the libwmgeneral folder.

Signed-off-by: Rodolfo García Peñas (kix) <kix@kix.es>