libdockapp: Make DAParseArguments() optional.
Some dockapps may want to handle command line options themselves, so we
make this function optional.

Previously, if this function was skipped, then a segfault would result when
trying to access the _daContext global while first creating the dockapp
window.

Now we check if _daContext has been initialized first, and if not, we
initialize it.