libdockapp: Add DAMakeShapeFromData() and DAMakeShapeFromFile() functions.
libdockapp supports shaped dockapps with the DASetShape() function, but this
function requires as input a bitmap.  Previously, there was no support for
creating such a bitmap from XBM data without using Xlib directly.

We add two functions, DAMakeShapeFromData(), which is a wrapper around
XCreateBitmapFromData and allows developers to #include XBM data, and
DAMakeShapeFromFile(), which is a wrapper around XReadBitmapfile and
lets developers specify the path to an XBM file.