fookb: Remove --enable-wmaker compile-time option.
Now that we use libdockapp's --windowed option to allow users to choose
whether to run fookb as a dockapp or as a normal windowed application at
runtime, the --enable-wmaker option compile time option did very little.

In particular, it used a config file in ~/GNUstep/Defaults and it printed
a warning message when users used large icon files.

We now use the ~/.fookb config file in all instances and always print the
warning message.