wmmon: Only compute vertical position of hint lines once.
The system load monitor displays green hint lines at every 100 units once
load average rises above 100.

Previously, the vertical position of these lines was computed before
each and every pixel was drawn.  But since the horizontal position
isn't needed for this computation, we move it up a level to the
outer for loop.