wmmon: Add support for larger load averages.
Previously, a green hint line was drawn on the CPU window for each
multiple of 1.  This meant as the load average approached 40 (the
height of the window), the window would gradually fill up entirely
with green.

To get around this, we switch to a yellow hint line for each
multiple of 10 once the load passes 10.  And then a red hint line
for each multiple of 100 once the load passes 100.

This fixes Debian bug #894801, reported by Pedro Gimeno
<debian.bts@personal.formauri.es>.