Fix multiple definitions of display variable.
The display variable is declared in a header with no explicit linkage.
This results in there being multiple definitions of it in wmbiff.o and
libwmgeneral.a and a failure during linking with gcc 10.

Add `extern` to the header declaration and a separate declaration with no
linkage in wmgeneral.c where it is assigned.

Link: https://bugs.debian.org/957937
Signed-off-by: Jeremy Sowden <jeremy@azazel.net>