wmix: fix multiple definitions of global variables.
The mixer API variables are declared in a header with no explicit
linkage.  This results in there being definitions of them in multiple
object files, which causes a link failure with GCC 10, since this
uses -fno-common by default.

Add `extern` to the header declarations and separate declarations with
no linkage in mixer.c where they are assigned.

Link: https://bugs.debian.org/957947
Signed-off-by: Jeremy Sowden <jeremy@azazel.net>