wmmoonclock: fix multiple definitions of some global variables.
A number of variables are declared in a header with no explicit linkage.
This results in there being definitions of them in multiple object
files, which causes a link failure with GCC 10, since this uses
-fno-common by default.

Add `extern` to the header declarations and separate declarations with
no linkage in xutils.c where they are assigned.

Link: https://bugs.debian.org/957950
Signed-off-by: Jeremy Sowden <jeremy@azazel.net>