wmacpi: fix multiple definitions of two global variables.
A couple of variables are declared in a header with no explicit linkage.
This results in there being definitions of them in multiple object
files, which causes a link failure with GCC 10, since this uses
-fno-common by default.

Add `extern` to the header declarations and separate declarations with
no linkage in libacpi.c where they are initialized.

Link: https://bugs.debian.org/957938
Signed-off-by: Jeremy Sowden <jeremy@azazel.net>