libdockapp: fix multiple definitions of global variable.
The `display` variable is declared in wmgeneral.h with no explicit
linkage.  This may result in there being definitions of it in the
library and the object-files of applications which link against it,
which causes link failures when building these applications with GCC 10,
since this uses -fno-common by default.

Add `extern` to the header declaration and a separate declaration with
no linkage in wmgeneral.c where it is initialized.

Signed-off-by: Jeremy Sowden <jeremy@azazel.net>