wmbattery: avoid truncating ACPI file paths
Instead of writing the static `SYSFS_PATH` string to a buffer, and
passing the buffer to `opendir`, remove the buffer and pass the static
string directly.

Instead of using a fixed size for the ACPI file path buffer, size it
based on the sizes of the arguments that snprintf will write to it.

Signed-off-by: Jeremy Sowden <jeremy@azazel.net>